+++ src/boot/AppEventHandlers.js
@@ -9,7 +9,8 @@ import * as ScreenOrientation from 'expo-screen-orientation';
 import type { GlobalDispatch, Orientation as OrientationT } from '../types';
 import { createStyleSheet } from '../styles';
 import { connectGlobal } from '../react-redux';
-import { handleInitialNotification, NotificationListener } from '../notification';
+import { handleInitialNotification } from '../notification';
+import NotificationListener from '../notification/NotificationListener';
 import { ShareReceivedListener, handleInitialShare } from '../sharing';
 import { appOnline, appOrientation } from '../actions';
 import PresenceHeartbeat from '../presence/PresenceHeartbeat';
