+++ src/notification/index.js
@@ -1,12 +1,11 @@
 /* @flow strict-local */
-import { DeviceEventEmitter, NativeModules, Platform } from 'react-native';
+import { NativeModules, Platform } from 'react-native';
 import PushNotificationIOS from '@react-native-community/push-notification-ios';
-import type { PushNotificationEventName } from '@react-native-community/push-notification-ios';
 
 import type { Notification } from './types';
 import type { GlobalDispatch, Account, Narrow, Stream, UserId, UserOrBot } from '../types';
 import { topicNarrow, pm1to1NarrowFromUser, pmNarrowFromRecipients } from '../utils/narrow';
-import type { JSONable, JSONableDict } from '../utils/jsonable';
+import type { JSONable } from '../utils/jsonable';
 import * as logging from '../utils/logging';
 import { gotPushToken, sendAllPushToken, narrowToNotification } from './notificationActions';
 import { fromPushNotificationIOS } from './extract';
@@ -201,7 +200,7 @@ export const handleInitialNotification = async (dispatch: GlobalDispatch) => {
  *
  * Returns null (and logs a warning or error) if getting the token failed.
  */
-const androidGetToken = async (dispatch: GlobalDispatch): Promise<mixed> => {
+export const androidGetToken = async (dispatch: GlobalDispatch): Promise<mixed> => {
   try {
     return await NativeModules.Notifications.getToken();
   } catch (e) {
@@ -228,145 +227,6 @@ const androidGetToken = async (dispatch: GlobalDispatch): Promise<mixed> => {
     return null;
   }
 };
-
-/**
- * From ios/RNCPushNotificationIOS.m in @rnc/push-notification-ios at 1.2.2.
- */
-type NotificationRegistrationFailedEvent = {|
-  // NSError.localizedDescription, see
-  // https://developer.apple.com/documentation/foundation/nserror/1414418-localizeddescription
-  message: string,
-
-  // NSError.code, see
-  // https://developer.apple.com/documentation/foundation/nserror/1409165-code
-  code: number,
-
-  // NSError.userInfo, see
-  // https://developer.apple.com/documentation/foundation/nserror/1411580-userinfo
-  details: JSONableDict,
-|};
-
-/**
- * Listens for notification-related events.
- *
- * An instance of this doesn't affect the subscriptions of any other
- * instance, or anything else.
- */
-export class NotificationListener {
-  dispatch: GlobalDispatch;
-  unsubs: Array<() => void> = [];
-
-  constructor(dispatch: GlobalDispatch) {
-    this.dispatch = dispatch;
-  }
-
-  /** Private. */
-  listenIOS(name: PushNotificationEventName, handler: (...empty) => void | Promise<void>) {
-    // In the native code, the PushNotificationEventName we pass here
-    // is mapped to something else (see implementation):
-    //
-    // 'notification'      -> 'remoteNotificationReceived'
-    // 'localNotification' -> 'localNotificationReceived'
-    // 'register'          -> 'remoteNotificationsRegistered'
-    // 'registrationError' -> 'remoteNotificationRegistrationError'
-    PushNotificationIOS.addEventListener(name, handler);
-    this.unsubs.push(() => PushNotificationIOS.removeEventListener(name));
-  }
-
-  /** Private. */
-  listenAndroid(name: string, handler: (...empty) => void | Promise<void>) {
-    const subscription = DeviceEventEmitter.addListener(name, handler);
-    this.unsubs.push(() => subscription.remove());
-  }
-
-  /** Private. */
-  unlistenAll() {
-    while (this.unsubs.length > 0) {
-      this.unsubs.pop()();
-    }
-  }
-
-  /** Private. */
-  handleNotificationOpen: Notification => void = notification => {
-    this.dispatch(narrowToNotification(notification));
-  };
-
-  /**
-   * Private.
-   *
-   * @param deviceToken This should be a `?string`, but there's no typechecking
-   *   at the registration site to allow us to ensure it. As we've been burned
-   *   by unexpected types here before, we do the validation explicitly.
-   */
-  handleDeviceToken: mixed => Promise<void> = async deviceToken => {
-    // Null device tokens are known to occur (at least) on Android emulators
-    // without Google Play services, and have been reported in other scenarios.
-    // See https://stackoverflow.com/q/37517860 for relevant discussion.
-    //
-    // Otherwise, a device token should be some (platform-dependent and largely
-    // unspecified) flavor of string.
-    if (deviceToken !== null && typeof deviceToken !== 'string') {
-      /* $FlowFixMe[incompatible-type]: `deviceToken` probably _is_
-         JSONable, but we can only hope. */
-      const token: JSONable = deviceToken;
-      logging.error('Received invalid device token', { token });
-      // Take no further action.
-      return;
-    }
-
-    this.dispatch(gotPushToken(deviceToken));
-    await this.dispatch(sendAllPushToken());
-  };
-
-  /** Private. */
-  handleIOSRegistrationFailure: NotificationRegistrationFailedEvent => void = err => {
-    logging.warn(`Failed to register iOS push token: ${err.code}`, {
-      raw_error: err,
-    });
-  };
-
-  /** Start listening.  Don't call twice without intervening `stop`. */
-  async start() {
-    if (Platform.OS === 'android') {
-      // On Android, the object passed to the handler is constructed in
-      // FcmMessage.kt, and will always be a Notification.
-      this.listenAndroid('notificationOpened', this.handleNotificationOpen);
-      this.listenAndroid('remoteNotificationsRegistered', this.handleDeviceToken);
-    } else {
-      this.listenIOS('notification', (notification: PushNotificationIOS) => {
-        const dataFromAPNs = fromPushNotificationIOS(notification);
-        if (!dataFromAPNs) {
-          return;
-        }
-        this.handleNotificationOpen(dataFromAPNs);
-      });
-      this.listenIOS('register', this.handleDeviceToken);
-      this.listenIOS('registrationError', this.handleIOSRegistrationFailure);
-    }
-
-    if (Platform.OS === 'android') {
-      // A bug was introduced in 3730be4c8 that delayed the setup of
-      // our listener for 'remoteNotificationsRegistered' until a time
-      // after the event was emitted from the native code. Until we
-      // settle on a better, more consistent architecture, just grab
-      // the token here and do the same thing our handler does (by
-      // just calling the handler).
-      const token = await androidGetToken(this.dispatch);
-      if (token !== null) {
-        this.handleDeviceToken(token);
-      }
-    } else {
-      // On iOS, we come back to this later: after the initial fetch, we
-      // end up calling `getNotificationToken`, below, and that will cause
-      // us to get the token if the user gives us notification permission.
-    }
-  }
-
-  /** Stop listening. */
-  stop() {
-    this.unlistenAll();
-  }
-}
 
 /**
  * Try to cause a `remoteNotificationsRegistered` event.
@@ -421,4 +281,31 @@ export const getNotificationToken = () => {
   } else {
     // On Android, we do this at application startup.
   }
+};
+
+/**
+ * Act on having learned a new device token.
+ *
+ * @param deviceToken This should be a `?string`, but there's no typechecking
+ *   at the registration site to allow us to ensure it. As we've been burned
+ *   by unexpected types here before, we do the validation explicitly.
+ */
+export const handleDeviceToken = async (deviceToken: mixed, dispatch: GlobalDispatch) => {
+  // Null device tokens are known to occur (at least) on Android emulators
+  // without Google Play services, and have been reported in other scenarios.
+  // See https://stackoverflow.com/q/37517860 for relevant discussion.
+  //
+  // Otherwise, a device token should be some (platform-dependent and largely
+  // unspecified) flavor of string.
+  if (deviceToken !== null && typeof deviceToken !== 'string') {
+    /* $FlowFixMe[incompatible-type]: `deviceToken` probably _is_
+         JSONable, but we can only hope. */
+    const token: JSONable = deviceToken;
+    logging.error('Received invalid device token', { token });
+    // Take no further action.
+    return;
+  }
+
+  dispatch(gotPushToken(deviceToken));
+  await dispatch(sendAllPushToken());
 };
